Move gid allocator from efs to lib; use in glusterfs
