+++ lib/gidallocator/allocator.go
@@ -24,7 +24,7 @@ import (
 	"sync"
 
 	"github.com/golang/glog"
-	"github.com/kubernetes-incubator/external-storage/gluster/glusterfs/pkg/allocator"
+	"github.com/kubernetes-incubator/external-storage/lib/allocator"
 	"github.com/kubernetes-incubator/external-storage/lib/controller"
 	"k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
